added the last of the tests... I think
